Starting up a bottoms-up PHP server-side implementation.
Just wanted to get a proof-of-concept going.  This currently fakes its
way through the authentication process, but there is no database storage
as of now so it's not practical yet.

Feel free to hack, slash, and delete at will.  =)  I've at least gotten
all the API error constants in and documented, and established a basic
dispatching method for processing requests.

Copy 2.0/, api/, and data/ directories to a PHP-capable server's root to
simulate the Last.FM experience.  The hosts file would need to alias
last.fm and ws.audioscrobbler.com, but I'm not bothering with that yet.