Indexes needed by track.scrobble,updatenowplaying
We dont use lower() in SELECT queries for 2.0 scrobbling as this will
eventually cause exceptions when doing INSERT without lower() because of
foreign key restrictions. We should do the same changes in gnukebox.
We should try to rewrite all queries that is using lower() so we
can remove the lowercase indexes.