User header
Rename maxiprofile.tpl to user-header.tpl
Instead of doing $smarty->assign(headerfile, '<header>.tpl') in php
files, we add the tpl file as a subheader parameter when including header.tpl
in page template.
It makes it easier to keep track of what templates are loaded when
modifying a theme.

Change header.tpl to also take a showbio parameter, if true it will
display a users bio.