Add gnufm_key to config
Add $gnufm_key = 'default_gnufm_32_char_identifier' to config.php,
and use it when getting a scrobble session id in login.php and when
scrobbling from the website player.
This lets us re-use the session id we set at login time, and avoid
creating another session id for the player.